+++ javaslang/src/main/java/javaslang/collection/HashSet.java
@@ -5,11 +5,17 @@
  */
 package javaslang.collection;
 
-import javaslang.*;
+import javaslang.Kind1;
+import javaslang.Tuple;
+import javaslang.Tuple2;
+import javaslang.Tuple3;
 import javaslang.control.Option;
 
 import java.io.*;
-import java.util.*;
+import java.util.ArrayList;
+import java.util.Comparator;
+import java.util.NoSuchElementException;
+import java.util.Objects;
 import java.util.function.*;
 import java.util.stream.Collector;
 
@@ -506,7 +512,7 @@ public final class HashSet<T> implements Kind1<HashSet<?>, T>, Set<T>, Serializa
     }
 
     @Override
-    public HashSet<T> drop(long n) {
+    public HashSet<T> drop(int n) {
         if (n <= 0) {
             return this;
         } else {
@@ -515,7 +521,7 @@ public final class HashSet<T> implements Kind1<HashSet<?>, T>, Set<T>, Serializa
     }
 
     @Override
-    public HashSet<T> dropRight(long n) {
+    public HashSet<T> dropRight(int n) {
         return drop(n);
     }
 
@@ -569,7 +575,7 @@ public final class HashSet<T> implements Kind1<HashSet<?>, T>, Set<T>, Serializa
     }
 
     @Override
-    public Iterator<HashSet<T>> grouped(long size) {
+    public Iterator<HashSet<T>> grouped(int size) {
         return sliding(size, size);
     }
 
@@ -723,12 +729,12 @@ public final class HashSet<T> implements Kind1<HashSet<?>, T>, Set<T>, Serializa
     }
 
     @Override
-    public Iterator<HashSet<T>> sliding(long size) {
+    public Iterator<HashSet<T>> sliding(int size) {
         return sliding(size, 1);
     }
 
     @Override
-    public Iterator<HashSet<T>> sliding(long size, long step) {
+    public Iterator<HashSet<T>> sliding(int size, int step) {
         return iterator().sliding(size, step).map(HashSet::ofAll);
     }
 
@@ -757,7 +763,7 @@ public final class HashSet<T> implements Kind1<HashSet<?>, T>, Set<T>, Serializa
     }
 
     @Override
-    public HashSet<T> take(long n) {
+    public HashSet<T> take(int n) {
         if (tree.size() <= n) {
             return this;
         }
@@ -765,7 +771,7 @@ public final class HashSet<T> implements Kind1<HashSet<?>, T>, Set<T>, Serializa
     }
 
     @Override
-    public HashSet<T> takeRight(long n) {
+    public HashSet<T> takeRight(int n) {
         return take(n);
     }
 
@@ -857,12 +863,12 @@ public final class HashSet<T> implements Kind1<HashSet<?>, T>, Set<T>, Serializa
     }
 
     @Override
-    public HashSet<Tuple2<T, Long>> zipWithIndex() {
+    public HashSet<Tuple2<T, Integer>> zipWithIndex() {
         return zipWithIndex(Tuple::of);
     }
 
     @Override
-    public <U> HashSet<U> zipWithIndex(BiFunction<? super T, ? super Long, ? extends U> mapper) {
+    public <U> HashSet<U> zipWithIndex(BiFunction<? super T, ? super Integer, ? extends U> mapper) {
         Objects.requireNonNull(mapper, "mapper is null");
         return HashSet.ofAll(iterator().zipWithIndex(mapper));
     }
