+++ javaslang/src/test/java/javaslang/collection/AbstractMultimapTest.java
@@ -607,13 +607,13 @@ public abstract class AbstractMultimapTest extends AbstractTraversableTest {
 
     @Test
     public void shouldZipNonNilWithIndex() {
-        final Seq<Tuple2<Tuple2<Integer, Integer>, Long>> actual = emptyIntInt()
+        final Seq<Tuple2<Tuple2<Integer, Integer>, Integer>> actual = emptyIntInt()
                 .put(0, 0)
                 .put(1, 1)
                 .put(2, 2)
                 .zipWithIndex();
         assertThat(actual).isEqualTo(
-                Stream.of(Tuple.of(Tuple.of(0, 0), 0L), Tuple.of(Tuple.of(1, 1), 1L), Tuple.of(Tuple.of(2, 2), 2L)));
+                Stream.of(Tuple.of(Tuple.of(0, 0), 0), Tuple.of(Tuple.of(1, 1), 1), Tuple.of(Tuple.of(2, 2), 2)));
     }
 
     // -- zipAll
